工厂迁移
